added check for existance of err file and comment cleanup

+++ pgsql-backup-all.sh
@@ -7,7 +7,7 @@
 # this is a daily cronjob
 CWD=/backup
 
-#set some variables for the log rotation
+# Set some variables for the log rotation
 curDate=`date "+%Y-%m-%d"`
 expDateFile=`date --date="last week" "+%Y-%m-%d"`
 allDB=`/usr/bin/psql -U postgres -d postgres -q -t -c 'SELECT datname from pg_database' | grep -v 'template0' | grep -v 'template1'`
@@ -27,11 +27,13 @@ for i in $allDB; do
   fi
 done
 
-#Delete file from a week ago
+# Delete file from a week ago
 if [ -d $CWD/daily/$HOSTNAME/$expDateFile ]; then
 	logger "pgsql-backup-all.sh: removing $CWD/daily/$HOSTNAME/$expDateFile"
         rm -rf $CWD/daily/$HOSTNAME/$expDateFile
 fi
 
-#cleanup err file
-rm /tmp/err
+# Cleanup files
+if [ -a /tmp/err ]; then
+  rm /tmp/err
+fi
